전체적으로 한번 다듬었습니다.
UIFrame 대신 Card라는 이름을 사용하도록 했습니다.
UIFrame이 default export 대신 named export를 사용하게 했습니다.
Welcome2 페이지에서 사용한 의존성 배열을 linter가 알려주는대로 변경했습니다.
게임 버튼을 통합했습니다. 추후 컴포넌트로 만들 수 있어보입니다.
