--- front/app/main/@main/CreateGameButton.tsx
@@ -1,15 +1,0 @@
-export function CreateGameButton() {
-    return (
-        <button
-            type="button"
-            // onClick={asyncRoute}
-            className="relative flex place-items-center"
-        >
-            <div className="gradient-border before:pointer-elvents-none group relative flex w-[262px] flex-col items-start gap-2 rounded-[28px] border-transparent bg-windowGlass/30 p-px px-5 py-4 backdrop-blur-[20px] backdrop-brightness-100 transition-colors before:absolute before:inset-0 before:rounded-[28px] before:p-px before:content-[''] hover:bg-primary focus:outline-2 focus:outline-gray-100 focus:ring focus:ring-violet-300 active:bg-violet-700 hover:dark:border-purple-700 hover:dark:bg-neutral-800/30">
-                <h2 className="text-sans text-2xl font-medium">
-                    Create Game
-                </h2>
-            </div>
-        </button>
-    );
-}
